debugging and refactoring
