--- server/model/user.ts
@@ -1,5 +1,0 @@
-export interface User {
-    user_id : number;
-    amount : number;
-}
-
