+++ server/src/index.ts
@@ -1,0 +1,20 @@
+import { app } from "./start";
+
+
+/**
+* App Variables
+*/
+
+
+const PORT : number = 8080;
+
+
+/**
+* Server Activation
+*/
+
+
+app.listen(PORT, () => {
+    console.log(`listening on port ${PORT}`);
+});
+
