+++ server/src/model/user.ts
@@ -1,0 +1,5 @@
+export interface User {
+    user_id : number;
+    amount : number;
+}
+
